Split the ProjectApp into smaller sub-apps
